♻️ refactor(pinctrl): 简化架构并统一 Iomux 定义
- 删除通用层 pinmux 和 fdt 模块,简化依赖关系
- 将 Iomux 定义为 bitflags 类型并移至 pinconf 模块
- 删除独立的 pinctrl.rs,直接在 mod.rs 中实现 Pinctrl
- 移除 Function 类型的公共导出,简化 API 表面
- 更新所有 Iomux 引用路径为 pinctrl::Iomux
- 清理 clock/mod.rs 中未使用的 RangeBounds 导入

这次重构简化了模块层次结构,减少了文件数量,并使 Iomux
定义更接近其使用位置,提高了代码的可维护性。
