+++ src/lib.rs
@@ -23,7 +23,7 @@ use core::ptr::NonNull;
 pub use variants::*;
 
 pub use pinctrl::id::*;
-pub use pinctrl::{Function, GpioDirection, PinConfig, PinctrlResult, Pull};
+pub use pinctrl::{GpioDirection, PinConfig, PinctrlResult, Pull};
 pub type Mmio = NonNull<u8>;
 pub use rst::{ResetRockchip, RstId};
